requirements.txt
appdirs==1.4.1
Chameleon==3.1
hupper==0.4.2
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==0.23
packaging==16.8
PasteDeploy==1.5.2
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.1.10
pyramid==1.8.2
pyramid-chameleon==0.3
pyramid-debugtoolbar==3.0.5
pyramid-jinja2==2.7
pyramid-jwt==1.1
pyramid-mako==1.0.2
repoze.lru==0.6
requests==2.13.0
six==1.10.0
translationstring==1.3
venusian==1.0
waitress==1.0.2
WebOb==1.7.1
zope.deprecation==4.2.0
zope.interface==4.3.3
